// src/@types/events.d.ts
// Generated by https://quicktype.io

export enum CountryAlpha3 {
  Esp = 'ESP',
  Fra = 'FRA',
  Gbr = 'GBR',
  Ita = 'ITA',
  Usa = 'USA'
}

export enum CountryID {
  The54D8A87238Fe5D27D500988A = '54d8a87238fe5d27d500988a',
  The5982Dffafa1614A04Fe33793 = '5982dffafa1614a04fe33793',
  The5B23E4B1F3C6455A193436C2 = '5b23e4b1f3c6455a193436c2',
  The5D68E1850E0D4E3Dd6E00B8D = '5d68e1850e0d4e3dd6e00b8d',
  The5D8Cee4B3142259150C465E1 = '5d8cee4b3142259150c465e1'
}

export enum CountryName {
  France = 'France',
  Italia = 'Italia',
  Spain = 'Spain',
  UnitedKingdom = 'United Kingdom',
  UnitedStatesOfAmerica = 'United States of America'
}

export enum Country {
  France = 'France',
  Italy = 'Italy',
  Spain = 'Spain',
  UnitedKingdom = 'United Kingdom',
  UnitedStates = 'United States'
}

export enum Timezone {
  AmericaLosAngeles = 'America/Los_Angeles',
  AmericaNewYork = 'America/New_York',
  EuropeLondon = 'Europe/London',
  EuropeMadrid = 'Europe/Madrid',
  EuropeParis = 'Europe/Paris',
  EuropeRome = 'Europe/Rome'
}

export enum Type {
  Event = 'event'
}

export enum Currency {
  Eur = 'EUR',
  Gbp = 'GBP',
  Usd = 'USD'
}

export enum Flag {
  BarCode = 'bar-code',
  Cancelled = 'cancelled',
  CodeLocked = 'code-locked',
  GenerateNewCodeOnTransfer = 'generate-new-code-on-transfer',
  GoingAhead = 'going_ahead',
  NoWaitingList = 'no-waiting-list',
  QrCode = 'qr-code'
}

export enum Status {
  Cancelled = 'cancelled',
  OffSale = 'off-sale',
  OnSale = 'on-sale'
}

export interface Track {
  open_url: string
  preview_url: null | string
  title: string
}

export interface City {
  code: string
  country_alpha3: CountryAlpha3
  country_id: CountryID
  country_name: CountryName
  id: string
  name: string
}

export interface DetailedArtist {
  headliner: boolean
  id: number
  name: string
}

export interface EventImages {
  brand: null
  landscape: string
  portrait: string
  square: string
}

export interface Lineup {
  details: string
  time: string
}

export interface Location {
  accuracy: number
  city: string
  country: Country
  lat: number
  lng: number
  place: Timezone
}

export interface Promoter {
  id: number
  name: string
}

export interface Price {
  face_value: number
  fees: number
  total: number
}

export interface TicketType {
  id: number
  name: string
  price: Price
  sold_out: boolean
}

export interface Venue {
  city: City
  id: number
  name: string
}

export interface Links {
  next: string
  self: string
}

export interface PageLinks {
  next: string | null
  self: string
}

export interface Event {
  raw_description: string
  sale_start_date: Date
  announcement_date: Date
  status: Status
  price: null
  tags: string[]
  show_price_breakdown: boolean
  name: string
  timezone: Timezone
  venues: Venue[]
  sale_end_date: Date
  type: Type
  destination_event_perm_name: null
  apple_music_tracks: Track[]
  perm_name: string
  id: string
  genre_tags: string[]
  spotify_tracks: Track[]
  detailed_artists: DetailedArtist[]
  event_images: EventImages
  artists: string[]
  cities: City[]
  location: Location
  checksum: string
  age_limit: string
  int_id: number
  presented_by: string
  linkout_type: null
  images: string[]
  destination_event_id: null
  sold_out: boolean
  hash: string
  date_end: Date
  date: Date
  ticket_types: TicketType[]
  type_tags: string[]
  flags: Flag[]
  lineup: Lineup[]
  links: string[]
  external_url: null
  description: string
  featured: boolean
  url: string
  address: string
  venue: string
  currency: Currency
  promoters: Promoter[]
}

export interface ApiData {
  data: Event[]
  links: PageLinks
}